Windows Boot: Fix MmSaveStateIoInfo struct layout and copy logic

+++ core/patina_internal_cpu/src/save_state.rs
@@ -275,7 +275,7 @@ pub struct ParsedIoInfo {
     /// Number of bytes transferred (1, 2, or 4).
     pub byte_count: usize,
     /// I/O port address.
-    pub io_port: u32,
+    pub io_port: u16,
 }
 
 /// `EFI_MM_SAVE_STATE_IO_INFO` — written to the user buffer when reading
@@ -286,7 +286,7 @@ pub struct MmSaveStateIoInfo {
     /// I/O data value (from RAX, zero-extended).
     pub io_data: u64,
     /// I/O port address.
-    pub io_port: u64,
+    pub io_port: u16,
     /// I/O width enum (`EFI_MM_SAVE_STATE_IO_WIDTH`).
     pub io_width: u32,
     /// I/O type enum (`EFI_MM_SAVE_STATE_IO_TYPE`).
